improved navbar spaghetti code
